% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{col2hex}
\alias{col2hex}
\title{Convert color name to HEX}
\usage{
col2hex(color)
}
\arguments{
\item{color}{Vector with R-color names}
}
\value{
Character
}
\description{
Return a color in HEX format as a character given a R-color name as character.
}
\examples{
col2hex(c("lightblue", "aquamarine4"))
}
