% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.r
\name{crabtree.colors}
\alias{crabtree.colors}
\title{Designer and university colors}
\usage{
crabtree.colors(set = c("nu", "psu", "cu", "um", "phistoric"))
}
\arguments{
\item{set}{Character string denoting a color palette.}
}
\value{
Vector of character strings from the selected color palette.
}
\description{
Function returns a vector of character strings from a selected color palette.
}
\examples{
\dontrun{
# University of Michigan colors
(pal = crabtree.colors("um"))
pie(rep(1, length(pal)), labels = sprintf("\%d (\%s)", seq_along(pal), pal), col = pal)

# Northwestern University colors
(pal = crabtree.colors("nu"))
pie(rep(1, length(pal)), labels = sprintf("\%d (\%s)", seq_along(pal), pal), col = pal)
}
}
\author{
Charles Crabtree \email{ccrabtr@umich.edu}
}
