% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{hex2col}
\alias{hex2col}
\title{Convert HEX to color name}
\usage{
hex2col(color, distinct = TRUE)
}
\arguments{
\item{color}{A character, color formatted as '#______' in HEX code}

\item{distinct}{Logical, whether the result should be one distinct name or more
(see 'Details'). (default: \code{TRUE})}
}
\value{
character
}
\description{
Return a R-color name as a character given a color in HEX format as character.
}
\details{
In R one color can have multiple names, e.g. 'white' and 'grey100'.
By setting distinct to \code{FALSE}, the function returns all possible names.
}
\examples{
hex2col("#458b74")
}
