% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{hex2rgb}
\alias{hex2rgb}
\title{Convert HEX to RGB}
\usage{
hex2rgb(color, arithmetic = FALSE)
}
\arguments{
\item{color}{A character, color formatted as '#______' in HEX code}

\item{arithmetic}{Logical, whether to use arithmetic scaling for the result.
See 'Details'.(default: \code{FALSE})}
}
\value{
Numeric vector of length 3.
}
\description{
Return a color in RGB format given its representation in HEX format as a character.
}
\details{
Two formats for the RGB values are implemented which can be accessed with
the \code{arithmetic} argument.
By default (i.e if \code{arithmetic} is set to \code{FALSE}), the range for every
color goes from 0 to 255.
Otherwise (i.e. if \code{arithmetic} is set to \code{TRUE}), the values are interpreted
as percentages given as numbers between 0 and 1.
}
\examples{
hex2rgb("#82AC85")
}
