% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_coloR.R
\name{show_colorR}
\alias{show_colorR}
\title{Show color}
\usage{
show_colorR(color, text_color = "black")
}
\arguments{
\item{color}{character or vector length 3, color in R-name, HEX or vector length 3}

\item{text_color}{character or vector length 3, color in R-name, HEX or vector length 3,
default: 'black'.}
}
\value{
plot
}
\description{
Given a color in format R-name as character, HEX as character or RGB as vector length 3,
return a rectangular plot of the color with the argument written in the middle.
}
\examples{
show_colorR("lightblue", text_color = "#AC5798")
}
