% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorR-package.R
\docType{package}
\name{colorR-package}
\alias{colorR}
\alias{colorR-package}
\title{colorR: Working with diffrent color formats in R}
\description{
Convert different formats for colors: rgb to hex and vice versa, color name in R to hex/rgb, and vice versa. Mixing to colors into one (rgb, hex, and color name allowed). Showing the color in a rectangular plot.
}
\author{
\strong{Maintainer}: Marleen Fiederer \email{fiederer.marleen@gmail.com}

}
\keyword{internal}
