% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixer.R
\name{mix}
\alias{mix}
\title{Mix to colors}
\usage{
mix(color1, color2, arithmetic = FALSE, as_HEX = FALSE)
}
\arguments{
\item{color1}{Character or numeric vector of length 3, the color as R-name, HEX or RGB.}

\item{color2}{Character or numeric vector of length 3, the color as R-name, HEX or RGB.}

\item{arithmetic}{Logical, whether to use arithmetic scaling. See 'Details'.
(default: \code{FALSE})}

\item{as_HEX}{Logical, whether to return the result in HEX or RGB format.}
}
\value{
Character if as_HEX=\code{TRUE}, vector of length 3 if as_HEX=\code{FALSE}.
}
\description{
Return color in HEX or RGB format given two colors which can be in HEX format
as character, R-name also as character or RGB format as a numeric vector of length 3.
The result is returned in HEX format as a character or in RGB format as a vector of length 3.
}
\details{
Two formats for the RGB values are implemented which can be accessed with
the \code{arithmetic} argument.
By default (i.e if \code{arithmetic} is set to \code{FALSE}), the range for every
color goes from 0 to 255.
Otherwise (i.e. if \code{arithmetic} is set to \code{TRUE}), the values are interpreted
as percentages given as numbers between 0 and 1.
}
\examples{
mix("aquamarine4", "#FC6C85")
}
