% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_coloR.R
\name{show_colorR}
\alias{show_colorR}
\title{Show color}
\usage{
show_colorR(color, text_color = "black")
}
\arguments{
\item{color}{Character or numeric vector of length 3, the color as R-name, HEX or RGB.}

\item{text_color}{Character or numeric vector of length 3, the text color as R-name, HEX or RGB.
(default: 'black')}
}
\value{
A plot as described above.
}
\description{
Given a color by its R-name as a character, in HEX format as character or RGB format as a vector of length 3,
return a rectangular plot of the color with the color written in the middle in the same format as the argument.
}
\examples{
show_colorR("lightblue", text_color = "#AC5798")
}
